requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
redis==5.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
APScheduler==3.10.4
Werkzeug==3.0.1
reportlab==4.0.7
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.24.3
opencv-python-headless==4.8.1.78
